requirements.txt
asgiref==3.9.1
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distlib==0.4.0
Django==5.2.5
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.18.0
fonttools==4.57.0
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.2.1
mysqlclient==2.2.7
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.1.0
qrcode==8.1
regex==2024.11.6
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uv==0.8.12
virtualenv==20.32.0
wcwidth==0.2.13
